renamed rule in readme(fixed....)
