+++ packages/eslint-plugin-pf-codemods/index.js
@@ -41,8 +41,8 @@ const rules = {
   "page-header-move-avatar": require('./lib/rules/page-header-move-avatar'),
   "refactor-breakpointmods": require('./lib/rules/refactor-breakpointmods'),
   "datatoolbar-rename-toolbar": require('./lib/rules/datatoolbar-rename-toolbar'),
-  "chartVoronoiContainer-remove-allowTooltip": require('./lib/rules/chart-remove-allowZoom'),
-  "chart-remove-allowZoom": require('./lib/rules/chart-remove-allowZoom'),
+  "chartVoronoiContainer-remove-allowTooltip": require('./lib/rules/chartVoronoiContainer-remove-allowTooltip'),
+  "chart-remove-allowZoom": require('./lib/rules/chartVoronoiContainer-remove-allowTooltip'),
 };
 
 module.exports = {
