+++ packages/eslint-plugin-pf-codemods/lib/rules/chartVoronoiContainer-remove-allowTooltip.js
@@ -1,0 +1,12 @@
+const { renameProps } = require('../helpers');
+
+// https://github.com/patternfly/patternfly-react/pull/4330
+const renames = {
+  'ChartVoronoiContainer': {
+    allowTooltip: ''
+  }
+};
+
+module.exports = {
+  create: renameProps(renames, '@patternfly/react-charts')
+};
