+++ packages/eslint-plugin-pf-codemods/test/rules/chart-remove-allowZoom.js
@@ -1,20 +1,52 @@
 const ruleTester = require('../ruletester');
 const rule = require('../../lib/rules/chart-remove-allowZoom');
 
-ruleTester.run("chartVoronoiContainer-remove-allowTooltip", rule, {
+ruleTester.run("createContainer-insteadof-allowZoom", rule, {
   valid: [
     {
-      code: `import { ChartVoronoiContainer } from '@patternfly/react-charts'; <ChartVoronoiContainer/>`,
+      code: `import { VictoryZoomContainer } from 'victory-zoom-container';
+import { Chart } from '@patternfly/react-charts';
+import { ChartAxis } from '@patternfly/react-charts';
+import { ChartGroup } from '@patternfly/react-charts';
+<Chart containerComponent={<VictoryZoomContainer />}>
+    <ChartAxis />
+        <ChartGroup offset={11} horizontal>
+            "some-chart"
+        </ChartGroup>
+</Chart>`,
     }
   ],
   invalid: [
     {
-      code: `import { ChartVoronoiContainer } from '@patternfly/react-charts'; <ChartVoronoiContainer allowTooltip={!allHidden} />`,
-      output: `import { ChartVoronoiContainer } from '@patternfly/react-charts'; <ChartVoronoiContainer  />`,
+      code: `import { Chart, ChartAxis, ChartGroup } from '@patternfly/react-charts';
+<Chart allowZoom>
+    <ChartAxis/>
+        <ChartGroup allowZoom offset={11} horizontal>
+            "some-chart"
+        </ChartGroup>
+</Chart>`,
+      output: `import { VictoryZoomContainer } from 'victory-zoom-container';
+import { Chart, ChartAxis, ChartGroup } from '@patternfly/react-charts';
+<Chart containerComponent={<VictoryZoomContainer />}>
+    <ChartAxis/>
+        <ChartGroup  offset={11} horizontal>
+            "some-chart"
+        </ChartGroup>
+</Chart>`,
       errors: [{
-        message: `allowTooltip prop for ChartVoronoiContainer has been removed`,
+        messageId: "missingImportMsg",
+        data: {
+          missingImports: 'VictoryZoomContainer',
+          ensurePackage: 'victory-zoom-container'
+        }},
+        {
+        message: `allowZoom prop for Chart has been renamed to containerComponent={<VictoryZoomContainer />}`,
         type: "JSXOpeningElement",
-      }]
+        },
+        {
+          message: `allowZoom prop for ChartGroup has been removed`,
+          type: "JSXOpeningElement",
+        }]
     },
   ]
 });
